Added hasher function to instruction.

+++ src/day_15/hash.rs
@@ -1,5 +1,5 @@
-pub fn hash(s: &str) -> u32 {
-    let ascii = s.chars().map(|c| c as u32);
+pub fn hash(s: &str) -> u64 {
+    let ascii = s.chars().map(|c| c as u64);
 
     return ascii.fold(0, |hash, code| (hash + code) * 17 % 256);
 }
